+++ frontend/src/components/TodoDetail.js
@@ -1,6 +1,5 @@
-import React, { useState, useEffect, useCallback } from 'react';
-import { useParams, useNavigate } from 'react-router-dom';
-import { 
+import React from 'react';
+import {
   Typography,
   Button,
   Container,
@@ -10,107 +9,123 @@ import {
   CircularProgress,
   Fade
 } from '@mui/material';
-import { 
+import {
   ArrowBack,
 } from '@mui/icons-material';
-import * as api from '../api';
 
 import TodoDetailHeader from './TodoDetailHeader';
 import TodoDetailEdit from './TodoDetailEdit';
 import TodoDetailView from './TodoDetailView';
+import { useTodoDetail } from '../hooks/useTodoDetail';
 
 function TodoDetail() {
-  const { id } = useParams();
-  const navigate = useNavigate();
-  const [todo, setTodo] = useState(null);
-  const [loading, setLoading] = useState(true);
-  const [saving, setSaving] = useState(false);
-  const [error, setError] = useState(null);
-  const [isEditing, setIsEditing] = useState(false);
-  const [editForm, setEditForm] = useState({
-    title: '',
-    description: '',
-    status: 'pending'
-  });
+  const {
+    todo,
+    loading,
+    saving,
+    error,
+    isEditing,
+    editForm,
+    setError,
+    setEditForm,
+    //fetchTodoDetail,
+    handleEdit,
+    handleCancelEdit,
+    handleSaveEdit,
+    handleDelete,
+    navigate
+  } = useTodoDetail();
+  // const { id } = useParams();
+  // const navigate = useNavigate();
+  // const [todo, setTodo] = useState(null);
+  // const [loading, setLoading] = useState(true);
+  // const [saving, setSaving] = useState(false);
+  // const [error, setError] = useState(null);
+  // const [isEditing, setIsEditing] = useState(false);
+  // const [editForm, setEditForm] = useState({
+  //   title: '',
+  //   description: '',
+  //   status: 'pending'
+  // });
 
-  const fetchTodoDetail = useCallback(async () => {
-    try {
-      setLoading(true);
-      const response = await api.readTodo(id);
-      setTodo(response.data);
-      setEditForm({
-        title: response.data.title,
-        description: response.data.description || '',
-        status: response.data.status
-      });
-      setError(null);
-    } catch (error) {
-      console.error('Error fetching todo detail:', error);
-      setError('TODOの詳細を取得できませんでした。');
-    } finally {
-      setLoading(false);
-    }
-  }, [id]);
+  // const fetchTodoDetail = useCallback(async () => {
+  //   try {
+  //     setLoading(true);
+  //     const response = await api.readTodo(id);
+  //     setTodo(response.data);
+  //     setEditForm({
+  //       title: response.data.title,
+  //       description: response.data.description || '',
+  //       status: response.data.status
+  //     });
+  //     setError(null);
+  //   } catch (error) {
+  //     console.error('Error fetching todo detail:', error);
+  //     setError('TODOの詳細を取得できませんでした。');
+  //   } finally {
+  //     setLoading(false);
+  //   }
+  // }, [id]);
 
-  useEffect(() => {
-    fetchTodoDetail();
-  }, [fetchTodoDetail]);
+  // useEffect(() => {
+  //   fetchTodoDetail();
+  // }, [fetchTodoDetail]);
 
-  const handleEdit = () => {
-    setIsEditing(true);
-  };
+  // const handleEdit = () => {
+  //   setIsEditing(true);
+  // };
 
-  const handleCancelEdit = () => {
-    // 変更がある場合は確認ダイアログを表示
-    const hasChanges = 
-      editForm.title !== todo.title ||
-      editForm.description !== (todo.description || '') ||
-      editForm.status !== todo.status;
+  // const handleCancelEdit = () => {
+  //   // 変更がある場合は確認ダイアログを表示
+  //   const hasChanges = 
+  //     editForm.title !== todo.title ||
+  //     editForm.description !== (todo.description || '') ||
+  //     editForm.status !== todo.status;
 
-    if (hasChanges && !window.confirm('変更を破棄しますか？')) {
-      return;
-    }
+  //   if (hasChanges && !window.confirm('変更を破棄しますか？')) {
+  //     return;
+  //   }
 
-    setIsEditing(false);
-    setEditForm({
-      title: todo.title,
-      description: todo.description || '',
-      status: todo.status
-    });
-  };
+  //   setIsEditing(false);
+  //   setEditForm({
+  //     title: todo.title,
+  //     description: todo.description || '',
+  //     status: todo.status
+  //   });
+  // };
 
-  const handleSaveEdit = async (e) => {
-    e.preventDefault();
-    if (!editForm.title.trim()) {
-      setError('タイトルは必須です。');
-      return;
-    }
+  // const handleSaveEdit = async (e) => {
+  //   e.preventDefault();
+  //   if (!editForm.title.trim()) {
+  //     setError('タイトルは必須です。');
+  //     return;
+  //   }
 
-    try {
-      setSaving(true);
-      await api.updateTodo(id, editForm);
-      setTodo({ ...todo, ...editForm, updated_at: new Date().toISOString() });
-      setIsEditing(false);
-      setError(null);
-    } catch (error) {
-      console.error('Error updating todo:', error);
-      setError('TODOの更新に失敗しました。');
-    } finally {
-      setSaving(false);
-    }
-  };
+  //   try {
+  //     setSaving(true);
+  //     await api.updateTodo(id, editForm);
+  //     setTodo({ ...todo, ...editForm, updated_at: new Date().toISOString() });
+  //     setIsEditing(false);
+  //     setError(null);
+  //   } catch (error) {
+  //     console.error('Error updating todo:', error);
+  //     setError('TODOの更新に失敗しました。');
+  //   } finally {
+  //     setSaving(false);
+  //   }
+  // };
 
-  const handleDelete = async () => {
-    if (window.confirm('このTODOを削除しますか？')) {
-      try {
-        await api.deleteTodo(id);
-        navigate('/');
-      } catch (error) {
-        console.error('Error deleting todo:', error);
-        setError('TODOの削除に失敗しました。');
-      }
-    }
-  };
+  // const handleDelete = async () => {
+  //   if (window.confirm('このTODOを削除しますか？')) {
+  //     try {
+  //       await api.deleteTodo(id);
+  //       navigate('/');
+  //     } catch (error) {
+  //       console.error('Error deleting todo:', error);
+  //       setError('TODOの削除に失敗しました。');
+  //     }
+  //   }
+  // };
 
   if (loading) {
     return (
@@ -129,9 +144,9 @@ function TodoDetail() {
         <Alert severity="error" sx={{ mb: 2 }}>
           {error}
         </Alert>
-        <Button 
-          variant="contained" 
-          onClick={() => navigate('/')} 
+        <Button
+          variant="contained"
+          onClick={() => navigate('/')}
           startIcon={<ArrowBack />}
         >
           一覧に戻る
@@ -146,9 +161,9 @@ function TodoDetail() {
         <Alert severity="warning" sx={{ mb: 2 }}>
           TODOが見つかりませんでした。
         </Alert>
-        <Button 
-          variant="contained" 
-          onClick={() => navigate('/')} 
+        <Button
+          variant="contained"
+          onClick={() => navigate('/')}
           startIcon={<ArrowBack />}
         >
           一覧に戻る
@@ -161,11 +176,11 @@ function TodoDetail() {
     <Container maxWidth="md" sx={{ py: 4 }}>
       <Stack spacing={3}>
         {/* ヘッダーエリア */}
-        <TodoDetailHeader 
-          navigate={navigate} 
-          isEditing={isEditing} 
-          handleEdit={handleEdit} 
-          handleDelete={handleDelete} 
+        <TodoDetailHeader
+          navigate={navigate}
+          isEditing={isEditing}
+          handleEdit={handleEdit}
+          handleDelete={handleDelete}
         />
 
         {/* エラー表示 */}
@@ -176,12 +191,12 @@ function TodoDetail() {
             </Alert>
           </Fade>
         )}
-        
+
         {/* メインコンテンツ */}
-        <Paper 
-          elevation={isEditing ? 4 : 2} 
-          sx={{ 
-            p: 4, 
+        <Paper
+          elevation={isEditing ? 4 : 2}
+          sx={{
+            p: 4,
             backgroundColor: isEditing ? '#fafafa' : 'white',
             border: isEditing ? '2px solid #1976d2' : 'none',
             transition: 'all 0.3s ease'
@@ -189,13 +204,13 @@ function TodoDetail() {
         >
           {isEditing ? (
             // 編集モード
-              <TodoDetailEdit 
-                editForm={editForm} 
-                saving={saving}
-                setEditForm={setEditForm} 
-                handleSaveEdit={handleSaveEdit}
-                handleCancelEdit={handleCancelEdit}
-              />
+            <TodoDetailEdit
+              editForm={editForm}
+              saving={saving}
+              setEditForm={setEditForm}
+              handleSaveEdit={handleSaveEdit}
+              handleCancelEdit={handleCancelEdit}
+            />
           ) : (
             // 表示モード
             <TodoDetailView todo={todo} />
